refactor(focusEditor): moved to utils.ts and removed repetitive code
